requirements.txt
RPi.GPIO
notify-run